Added an operators class, updated drivetrain, added AUTOmain, and added method to imu sensor
Created a simple linearopmode autonomous, updated drivetrain to actually work, it uses a rotation matrix to rotate the vector of the right stick by the yaw angle of the robot. Updated the main testing teleop to call reset_yaw() method, added reset_yaw() method to imu sensor class, and added operators class with a clamp function.