Refactor subsystem initialization and add TelemetryManager and fix creation of objects
Refactored Drivetrain and IMUSensor to accept RobotHardware instead of HardwareMap for improved dependency management. Introduced TelemetryManager for centralized telemetry logging. Moved and updated Sorter and Sorter_Automation classes to better reflect their subsystem and OpMode roles. Updated TeleOpMain and MotorTrackingTeleOp to use the new initialization patterns.

With the old old code, in each subsystem, we would create a new reference to what we wan. In the new code, we create all references in the teleop and pass those references into our subystem classes so we aren't creating repeats